requirements.txt
elasticsearch
requests
pytz
PyYAML